pyproject.toml
[project]
name = "pysyslink_toolkit"
version = "0.2.0"
description = "Toolkit for PySysLink block-based dynamic system simulator"
authors = [
    { name = "Pello Usabiaga", email = "pellousabiaga@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
  "PyYAML>6.0.1",
  "matplotlib",
  "mpld3",
  "dacite>1.9.2"
]


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
test = ["pytest"]